requirements.txt
qdrant-client
torch
transformers
langchain
pandas
numpy
fastapi
uvicorn
langchain-core
langchain-community
langchain-openai
openai
tiktoken
python-dotenv
